moving map state stuff into folder
