+++ Source/CovidAndWeatherVisualization/CovidAndWeatherVisualization/ClientApp/src/app/state/index.ts
@@ -1,1 +1,1 @@
-export * from './map-state.service';
+export * from './map/map-state.service';
