+++ Source/CovidAndWeatherVisualization/CovidAndWeatherVisualization/ClientApp/src/app/state/map/map-state.service.spec.ts
